[add] adjust

+++ ViewModel/DateBalanceModel.cs
@@ -5,9 +5,9 @@ namespace CloseFriendMyanamr.ViewModel
     public class DateBalanceModel
     {
         public DateTime Date { get; set; }
-        public double CashOpeningBalance { get; set; }
-        public double CashClosingBalance { get; set; }
-        public double BankOpeningBalance { get; set; }
-        public double BankClosingBalance { get; set; }
+        public decimal CashOpeningBalance { get; set; }
+        public decimal CashClosingBalance { get; set; }
+        public decimal BankOpeningBalance { get; set; }
+        public decimal BankClosingBalance { get; set; }
     }
 }
